requirements.txt
django-mongoengine==0.5.4
redis==4.2.2
django-redis==5.2.0
django-silk==4.2.0
django-cors-headers==3.5.0
gunicorn==20.1.0
mysqlclient==2.1.0
drf-yasg==1.20.0
gunicorn==20.1.0
django-cors-headers==3.5.0
Django==3.2